Kallos Claim Expansions & JCCS Modification
JCCS Modification permitted by shadedoom and included in this PR in order to properly conform with new kallos borders (see: https://discord.com/channels/797908768297713722/805953175009624134/831817814309863434)

Kallos Claim Expansions, Armenia (The Empire) is entirely inactive and after several weeks after snitching to confirm activity, it was confirmed that no activity existed in the region. Unless any users can properly confirm and construct infrastructure on The Empires behalf, this land is to be immediately claimed by Kallos, as Kallos has plans and uses for this land that exceeds whatever The Empire had intended for it.